Documentation for the section<em>Projects</em> is more complete. The contextual menu options are explained.
You can rename open documents and games. For the user session, this is transparent. All parts of the UI are renamed. You can rename a folder, and all documents in the hierarchy are renamed as well. If other users have one of those documents or games opened, a popup will tell them the document has been renamed. The popup identifies the user who performed the rename operations. The new name of the document is shown as well. The user just have to close the document and reopen it using the new name. All modifications are preserved.



git-svn-id: https://solucorp.bolixo.org/repos/solucorp/bolixo/trunk@3945 c3086684-d568-4a45-a57d-89166b08296d