whiteboard new features:
You can type text that appears along with the selected elements. If more than one element is selected, the text is repeated. Editing is limited. You can enter character and use the backspace key.

You can move the associated text in the center of the element, on top, bottom, to the left or right of the element. You do so using the middle mouse button and the ctrl key. Using the mouse button repeatedly, you cycle through the five positions.



git-svn-id: https://solucorp.bolixo.org/repos/solucorp/bolixo/trunk@3969 c3086684-d568-4a45-a57d-89166b08296d