Whiteboard: Various enhancements. Almost usable.
The bar menu is made of icons instead of single letters.

Independent lines may be added, moved and modified on the board.

Dash and dotted line types have been added.

Box types have been added. A button in the menu bar cycles through the various box type of the selected elements. The box types are solid, dash and dotted.



git-svn-id: https://solucorp.bolixo.org/repos/solucorp/bolixo/trunk@3970 c3086684-d568-4a45-a57d-89166b08296d