whiteboard: many enhancements.
You can place bullets text.

You can duplicate any selected elements just by hitting the enter key. The element will be unselected and a new element, selected, will appear under it. By hitting enter repeatedly, you can make many duplications. This is handy while you enter bullets.

Menu bar entries have been added to duplicate what is doable with mouse button and keyboard combination. You can modify the line type connecting a parent and a child. This is equivalent to the middle mouse button. You can move the text of an element. This is equivalent to ctrl+middle button.

When moving an element around its parent, the arrow is adjusted so it connect to the proper side of each element (bottom to top, right to left, ...). If many elements are connected to a parent, arrows are automatically moved so they do not cross each other.

Backspace on text supports UTF8.



git-svn-id: https://solucorp.bolixo.org/repos/solucorp/bolixo/trunk@3971 c3086684-d568-4a45-a57d-89166b08296d