Document embedding is getting closer to usefulness.
At this point the word processor can embed whiteboard views (only view A1 for now). The word processor can embed multiple whiteboard including multiple views of the same whiteboard.

(Given that the word processor is not ready for prime time, this is not so useful :- ( ).

Currently, embedding means that content is shown in the container document and updated in real time. You can move the embedded document around and scale it, but you can't modify the embedded document from the container document yet.



git-svn-id: https://solucorp.bolixo.org/repos/solucorp/bolixo/trunk@4002 c3086684-d568-4a45-a57d-89166b08296d