Some enhancements to the whiteboard.
When you delete a text bullet, the bullets below with the same X coordinate are raised. So it behaves like a word processor.

When you duplicate a text bullet (using enter), it creates a new one below, without text. This is a change. Previously, the text was also duplicated. It was not useful. Also, bullets below with the same X coordinates are moved down.

When moving an item, we wait 200 milliseconds after the item was selected to perform the move. This solves an annoying problem. Previously, selecting an item had the side effect of moving it slightly.


git-svn-id: https://solucorp.bolixo.org/repos/solucorp/bolixo/trunk@4131 c3086684-d568-4a45-a57d-89166b08296d