Bolixo now supports 'quoted replies'. This is an old concept, very useful, but lost for some reason. When replying to a message, you use the "quoted reply" in the context menu. The message is copied into the form, inserting a > in front of each line/paragraph. You remove the stuff you don't want and keep what is relevant. You place your reply just under. Any reader will immediately understand the context of your message.
The preview of the message correctly handles long paragraphs (quoted reply). The same new code was reused to support bullets and numbered lists. If a line starts with a star, it is an item from a bullet list. If it starts with a pound, it is a numbered list. You can put  up to 3 stars (our pounds) at the beginning of a line to create a sub or sub-sub bullet list.

 * item1
 * item2
 ** sub item1
 ** Sub item2



git-svn-id: https://solucorp.bolixo.org/repos/solucorp/bolixo/trunk@4753 c3086684-d568-4a45-a57d-89166b08296d